app/requirements.txt
googlemaps==4.10.0
kagglehub==0.3.6
mysql-connector-python==9.1.0
numpy==2.2.1
pandas==2.2.3
